Move gigachannel community to the corresponding component
