+++ experiment/popularity_community/initial_filling.py
@@ -21,7 +21,7 @@ from tribler_core.components.restapi import RESTComponent
 from tribler_core.components.socks_configurator import SocksServersComponent
 from tribler_core.components.torrent_checker import TorrentCheckerComponent
 from tribler_core.config.tribler_config import TriblerConfig
-from tribler_core.components.metadata_store.community.sync_strategy import RemovePeers
+from tribler_core.components.gigachannel.community.sync_strategy import RemovePeers
 from tribler_core.modules.popularity.community import PopularityCommunity
 from tribler_core.utilities.tiny_tribler_service import TinyTriblerService
