+++ src/seedbox/disseminator.py
@@ -30,7 +30,7 @@ from pony.orm import db_session
 import sentry_sdk
 
 from tribler_core.modules.libtorrent.torrentdef import TorrentDef
-from tribler_core.components.metadata_store.community import GigaChannelCommunity
+from tribler_core.components.gigachannel.community import GigaChannelCommunity
 from tribler_core.components.metadata_store.db.orm_bindings.channel_node import NEW
 from tribler_core.utilities.tiny_tribler_service import TinyTriblerService
