+++ src/tribler-core/tribler_core/components/gigachannel/community/settings.py
