+++ src/tribler-core/tribler_core/components/gigachannel/community/tests/test_gigachannel_community.py
@@ -11,12 +11,12 @@ from pony.orm import db_session
 
 import pytest
 
-from tribler_core.components.metadata_store.community.gigachannel_community import (
+from tribler_core.components.gigachannel.community.gigachannel_community import (
     ChannelsPeersMapping,
     GigaChannelCommunity,
     NoChannelSourcesException,
 )
-from tribler_core.components.metadata_store.community.settings import ChantSettings
+from tribler_core.components.gigachannel.community.settings import ChantSettings
 from tribler_core.components.metadata_store.db.store import MetadataStore
 from tribler_core.components.metadata_store.utils import RequestTimeoutException
 from tribler_core.modules.remote_query_community.settings import RemoteQueryCommunitySettings
