+++ src/tribler-core/tribler_core/components/gigachannel/gigachannel_component.py
@@ -2,12 +2,12 @@ from ipv8.peerdiscovery.discovery import RandomWalk
 from ipv8_service import IPv8
 from tribler_core.components.base import Component
 from tribler_core.components.ipv8 import Ipv8Component
-from tribler_core.components.metadata_store.community.gigachannel_community import GigaChannelCommunity, \
+from tribler_core.components.gigachannel.community.gigachannel_community import GigaChannelCommunity, \
     GigaChannelTestnetCommunity
 from tribler_core.components.metadata_store.metadata_store_component import MetadataStoreComponent
 from tribler_core.components.reporter import ReporterComponent
 from tribler_core.components.restapi import RESTComponent
-from tribler_core.components.metadata_store.community.sync_strategy import RemovePeers
+from tribler_core.components.gigachannel.community.sync_strategy import RemovePeers
 from tribler_core.restapi.rest_manager import RESTManager
 
 INFINITE = -1
