+++ src/tribler-core/tribler_core/components/metadata_store/restapi/tests/test_channels_endpoint.py
@@ -15,7 +15,7 @@ import pytest
 from tribler_common.simpledefs import CHANNEL_STATE
 
 from tribler_core.modules.libtorrent.torrentdef import TorrentDef
-from tribler_core.components.metadata_store.community.gigachannel_community import NoChannelSourcesException
+from tribler_core.components.gigachannel.community.gigachannel_community import NoChannelSourcesException
 from tribler_core.components.metadata_store.restapi.channels_endpoint import ChannelsEndpoint
 from tribler_core.components.metadata_store.db.serialization import CHANNEL_TORRENT, COLLECTION_NODE, REGULAR_TORRENT
 from tribler_core.components.metadata_store.utils import RequestTimeoutException
