+++ src/tribler-core/tribler_core/components/tests/test_tribler_components.py
@@ -4,7 +4,7 @@ import pytest
 
 from tribler_core.components.bandwidth_accounting.bandwidth_accounting_component import BandwidthAccountingComponent
 from tribler_core.components.base import Session, SessionError
-from tribler_core.components.gigachannel import GigaChannelComponent
+from tribler_core.components.gigachannel.gigachannel_component import GigaChannelComponent
 from tribler_core.components.ipv8 import Ipv8Component
 from tribler_core.components.libtorrent import LibtorrentComponent
 from tribler_core.components.masterkey import MasterKeyComponent
@@ -60,22 +60,6 @@ async def test_masterkey_component(tribler_config):
         assert comp.keypair
 
         await session.shutdown()
-
-
-async def test_giga_channel_component(tribler_config):
-    components = [MetadataStoreComponent(), RESTComponent(), MasterKeyComponent(), Ipv8Component(),
-                  GigaChannelComponent()]
-    session = Session(tribler_config, components)
-    with session:
-        comp = GigaChannelComponent.instance()
-        with patch.object(RESTManager, 'get_endpoint'):
-            await session.start()
-
-            assert comp.community
-            assert comp._rest_manager
-            assert comp._ipv8
-
-            await session.shutdown()
 
 
 async def test_ipv8_component(tribler_config):
